Refactor EventRule to make EventCall reusable
